+++ execution/order_test.go
@@ -219,7 +219,6 @@ func TestMarkPriceUpdateAfterPartialFill(t *testing.T) {
 	addAccount(tm, party1)
 	addAccount(tm, party2)
 	tm.broker.EXPECT().Send(gomock.Any()).AnyTimes()
-	tm.tradeStore.EXPECT().Add(gomock.Any()).AnyTimes()
 	tm.candleStore.EXPECT().AddTrade(gomock.Any()).AnyTimes()
 
 	orderBuy := &types.Order{
@@ -271,11 +270,11 @@ func TestExpireCancelGTCOrder(t *testing.T) {
 	tm := getTestMarket(t, now, closingAt)
 
 	addAccount(tm, party1)
-	tm.tradeStore.EXPECT().Add(gomock.Any()).AnyTimes()
 	tm.candleStore.EXPECT().AddTrade(gomock.Any()).AnyTimes()
 	tm.candleStore.EXPECT().Flush(gomock.Any(), gomock.Any()).AnyTimes()
 
 	orderBuy := &types.Order{
+		CreatedAt:   10000000000,
 		Status:      types.Order_STATUS_ACTIVE,
 		TimeInForce: types.Order_TIF_GTC,
 		Id:          "someid",
@@ -325,7 +324,6 @@ func TestAmendPartialFillCancelReplace(t *testing.T) {
 	addAccount(tm, party1)
 	addAccount(tm, party2)
 	tm.broker.EXPECT().Send(gomock.Any()).AnyTimes()
-	tm.tradeStore.EXPECT().Add(gomock.Any()).AnyTimes()
 	tm.candleStore.EXPECT().AddTrade(gomock.Any()).AnyTimes()
 	tm.candleStore.EXPECT().Flush(gomock.Any(), gomock.Any()).AnyTimes()