+++ gateway/graphql/resolvers.go
@@ -278,6 +278,13 @@ func (r *myQueryResolver) Parties(ctx context.Context, name *string) ([]*types.P
 	if err != nil {
 		return nil, err
 	}
+
+	// if we asked for a single party it may be null
+	// so then we return an empty slice
+	if party == nil {
+		return []*types.Party{}, nil
+	}
+
 	return []*types.Party{party}, nil
 }
 
@@ -511,7 +518,7 @@ func (r *myMarketResolver) Depth(ctx context.Context, market *Market, maxDepth *
 
 func (r *myMarketResolver) Candles(ctx context.Context, market *Market,
 	sinceRaw string, interval Interval) ([]*types.Candle, error) {
-	pinterval, err := convertInterval(interval)
+	pinterval, err := convertIntervalToProto(interval)
 	if err != nil {
 		r.log.Debug("interval convert error", logging.Error(err))
 	}
@@ -656,10 +663,9 @@ func (r *myPartyResolver) Orders(ctx context.Context, party *types.Party,
 
 	if len(res.Orders) > 0 {
 		return res.Orders, nil
-	} else {
-		// mandatory return field in schema
-		return []*types.Order{}, nil
 	}
+	// mandatory return field in schema
+	return []*types.Order{}, nil
 }
 
 func (r *myPartyResolver) Trades(ctx context.Context, party *types.Party,
@@ -685,10 +691,9 @@ func (r *myPartyResolver) Trades(ctx context.Context, party *types.Party,
 
 	if len(res.Trades) > 0 {
 		return res.Trades, nil
-	} else {
-		// mandatory return field in schema
-		return []*types.Trade{}, nil
 	}
+	// mandatory return field in schema
+	return []*types.Trade{}, nil
 }
 
 func (r *myPartyResolver) Positions(ctx context.Context, party *types.Party) ([]*types.Position, error) {
@@ -703,25 +708,9 @@ func (r *myPartyResolver) Positions(ctx context.Context, party *types.Party) ([]
 	}
 	if len(res.Positions) > 0 {
 		return res.Positions, nil
-	} else {
-		// mandatory return field in schema
-		return []*types.Position{}, nil
 	}
-}
-
-func AccountTypeToProto(acc AccountType) (types.AccountType, error) {
-	switch acc {
-	case AccountTypeGeneral:
-		return types.AccountType_ACCOUNT_TYPE_GENERAL, nil
-	case AccountTypeMargin:
-		return types.AccountType_ACCOUNT_TYPE_MARGIN, nil
-	case AccountTypeInsurance:
-		return types.AccountType_ACCOUNT_TYPE_INSURANCE, nil
-	case AccountTypeSettlement:
-		return types.AccountType_ACCOUNT_TYPE_SETTLEMENT, nil
-	default:
-		return types.AccountType_ACCOUNT_TYPE_UNSPECIFIED, fmt.Errorf("invalid account type %v, return default (ALL)", acc)
-	}
+	// mandatory return field in schema
+	return []*types.Position{}, nil
 }
 
 func (r *myPartyResolver) Accounts(ctx context.Context, party *types.Party,
@@ -743,7 +732,7 @@ func (r *myPartyResolver) Accounts(ctx context.Context, party *types.Party,
 		asst = *asset
 	}
 	if accType != nil {
-		accTy, err = AccountTypeToProto(*accType)
+		accTy, err = convertAccountTypeToProto(*accType)
 		if err != nil || (accTy != types.AccountType_ACCOUNT_TYPE_GENERAL && accTy != types.AccountType_ACCOUNT_TYPE_MARGIN) {
 			return nil, fmt.Errorf("invalid account type for party %v", accType)
 		}
@@ -767,10 +756,9 @@ func (r *myPartyResolver) Accounts(ctx context.Context, party *types.Party,
 
 	if len(res.Accounts) > 0 {
 		return res.Accounts, nil
-	} else {
-		// mandatory return field in schema
-		return []*types.Account{}, nil
 	}
+	// mandatory return field in schema
+	return []*types.Account{}, nil
 }
 
 func (r *myPartyResolver) Proposals(ctx context.Context, party *types.Party, inState *ProposalState) ([]*types.GovernanceData, error) {
@@ -838,7 +826,7 @@ func (r *myProposalResolver) State(ctx context.Context, data *types.GovernanceDa
 	if data == nil || data.Proposal == nil {
 		return "", ErrInvalidProposal
 	}
-	return ProposalStateFromProto(data.Proposal.State)
+	return convertProposalStateFromProto(data.Proposal.State)
 }
 
 func (r *myProposalResolver) Datetime(ctx context.Context, data *types.GovernanceData) (string, error) {
@@ -866,8 +854,12 @@ func (r *myProposalResolver) convertVotes(ctx context.Context, data []*types.Vot
 		if err != nil {
 			return nil, err
 		}
+		value, err := convertVoteValueFromProto(v.Value)
+		if err != nil {
+			return nil, err
+		}
 		result[i] = &Vote{
-			Value:    VoteValueFromProto(v.Value),
+			Value:    value,
 			Party:    voter,
 			Datetime: nanoTSToDatetime(v.Timestamp),
 		}
@@ -1057,48 +1049,11 @@ func (r *myMarketDepthResolver) Market(ctx context.Context, md *types.MarketDept
 
 type myOrderResolver VegaResolverRoot
 
-func RejectionReasonFromProtoOrderError(o types.OrderError) (RejectionReason, error) {
-	switch o {
-	case types.OrderError_ORDER_ERROR_INVALID_MARKET_ID:
-		return RejectionReasonInvalidMarketID, nil
-	case types.OrderError_ORDER_ERROR_INVALID_ORDER_ID:
-		return RejectionReasonInvalidOrderID, nil
-	case types.OrderError_ORDER_ERROR_OUT_OF_SEQUENCE:
-		return RejectionReasonOrderOutOfSequence, nil
-	case types.OrderError_ORDER_ERROR_INVALID_REMAINING_SIZE:
-		return RejectionReasonInvalidRemainingSize, nil
-	case types.OrderError_ORDER_ERROR_TIME_FAILURE:
-		return RejectionReasonTimeFailure, nil
-	case types.OrderError_ORDER_ERROR_REMOVAL_FAILURE:
-		return RejectionReasonOrderRemovalFailure, nil
-	case types.OrderError_ORDER_ERROR_INVALID_EXPIRATION_DATETIME:
-		return RejectionReasonInvalidExpirationTime, nil
-	case types.OrderError_ORDER_ERROR_INVALID_ORDER_REFERENCE:
-		return RejectionReasonInvalidOrderReference, nil
-	case types.OrderError_ORDER_ERROR_EDIT_NOT_ALLOWED:
-		return RejectionReasonEditNotAllowed, nil
-	case types.OrderError_ORDER_ERROR_AMEND_FAILURE:
-		return RejectionReasonOrderAmendFailure, nil
-	case types.OrderError_ORDER_ERROR_NOT_FOUND:
-		return RejectionReasonOrderNotFound, nil
-	case types.OrderError_ORDER_ERROR_INVALID_PARTY_ID:
-		return RejectionReasonInvalidPartyID, nil
-	case types.OrderError_ORDER_ERROR_MARKET_CLOSED:
-		return RejectionReasonMarketClosed, nil
-	case types.OrderError_ORDER_ERROR_MARGIN_CHECK_FAILED:
-		return RejectionReasonMarginCheckFailed, nil
-	case types.OrderError_ORDER_ERROR_INTERNAL_ERROR:
-		return RejectionReasonInternalError, nil
-	default:
-		return "", fmt.Errorf("invalid RejectionReason: %v", o)
-	}
-}
-
 func (r *myOrderResolver) RejectionReason(_ context.Context, o *types.Order) (*RejectionReason, error) {
 	if o.Reason == types.OrderError_ORDER_ERROR_NONE {
 		return nil, nil
 	}
-	reason, err := RejectionReasonFromProtoOrderError(o.Reason)
+	reason, err := convertRejectionReasonFromProto(o.Reason)
 	if err != nil {
 		return nil, err
 	}
@@ -1109,17 +1064,21 @@ func (r *myOrderResolver) Price(ctx context.Context, obj *types.Order) (string, 
 	return strconv.FormatUint(obj.Price, 10), nil
 }
 func (r *myOrderResolver) TimeInForce(ctx context.Context, obj *types.Order) (OrderTimeInForce, error) {
-	return OrderTimeInForce(obj.TimeInForce.String()), nil
+	return convertOrderTimeInForceFromProto(obj.TimeInForce)
 }
 
 func (r *myOrderResolver) Type(ctx context.Context, obj *types.Order) (*OrderType, error) {
-	t := OrderType(obj.Type.String())
+	t, err := convertOrderTypeFromProto(obj.Type)
+	if err != nil {
+		return nil, err
+	}
 	return &t, nil
 }
 
 func (r *myOrderResolver) Side(ctx context.Context, obj *types.Order) (Side, error) {
-	return Side(obj.Side.String()), nil
+	return convertSideFromProto(obj.Side)
 }
+
 func (r *myOrderResolver) Market(ctx context.Context, obj *types.Order) (*Market, error) {
 	if obj == nil {
 		return nil, errors.New("invalid order")
@@ -1138,9 +1097,11 @@ func (r *myOrderResolver) Size(ctx context.Context, obj *types.Order) (string, e
 func (r *myOrderResolver) Remaining(ctx context.Context, obj *types.Order) (string, error) {
 	return strconv.FormatUint(obj.Remaining, 10), nil
 }
+
 func (r *myOrderResolver) Status(ctx context.Context, obj *types.Order) (OrderStatus, error) {
-	return OrderStatus(obj.Status.String()), nil
+	return convertOrderStatusFromProto(obj.Status)
 }
+
 func (r *myOrderResolver) CreatedAt(ctx context.Context, obj *types.Order) (string, error) {
 	return vegatime.Format(vegatime.UnixNano(obj.CreatedAt)), nil
 }
@@ -1240,6 +1201,10 @@ func (r *myTradeResolver) Seller(ctx context.Context, obj *types.Trade) (*types.
 	return res.Party, nil
 }
 
+func (r *myTradeResolver) Type(ctx context.Context, obj *proto.Trade) (TradeType, error) {
+	return convertTradeTypeFromProto(obj.Type)
+}
+
 // END: Trade Resolver
 
 // BEGIN: Candle Resolver
@@ -1268,13 +1233,7 @@ func (r *myCandleResolver) Timestamp(ctx context.Context, obj *types.Candle) (st
 	return strconv.FormatInt(obj.Timestamp, 10), nil
 }
 func (r *myCandleResolver) Interval(ctx context.Context, obj *types.Candle) (Interval, error) {
-	interval := Interval(obj.Interval.String())
-	if interval.IsValid() {
-		return interval, nil
-	}
-	r.log.Warn("Interval conversion from proto to gql type failed, falling back to default: I15M",
-		logging.String("interval", interval.String()))
-	return IntervalI15m, nil
+	return convertIntervalFromProto(obj.Interval)
 }
 
 // END: Candle Resolver
@@ -1443,13 +1402,13 @@ func (r *myMutationResolver) PrepareOrderSubmit(ctx context.Context, market, par
 	}
 
 	order.PartyID = party
-	if order.TimeInForce, err = parseOrderTimeInForce(timeInForce); err != nil {
+	if order.TimeInForce, err = convertOrderTimeInForceToProto(timeInForce); err != nil {
 		return nil, err
 	}
-	if order.Side, err = parseSide(&side); err != nil {
+	if order.Side, err = convertSideToProto(side); err != nil {
 		return nil, err
 	}
-	if order.Type, err = parseOrderType(ty); err != nil {
+	if order.Type, err = convertOrderTypeToProto(ty); err != nil {
 		return nil, err
 	}
 
@@ -1518,11 +1477,6 @@ func (r *myMutationResolver) PrepareOrderCancel(ctx context.Context, id string, 
 
 func (r *myMutationResolver) PrepareProposal(
 	ctx context.Context, partyID string, reference *string, proposalTerms ProposalTermsInput) (*PreparedProposal, error) {
-
-	if proposalTerms.MinParticipationStake < 0 {
-		return nil, errors.New("minParticipationStake is invalid, must be positive")
-	}
-
 	var ref string
 	if reference != nil {
 		ref = *reference
@@ -1552,17 +1506,22 @@ func (r *myMutationResolver) PrepareVote(ctx context.Context, value VoteValue, p
 	if err != nil {
 		return nil, err
 	}
+	protoValue, err := convertVoteValueToProto(value)
+	if err != nil {
+		return nil, err
+	}
 	req := &protoapi.PrepareVoteRequest{
 		Vote: &types.Vote{
-			Value:      types.Vote_VALUE_NO,
+			Value:      protoValue,
 			PartyID:    partyID,
 			ProposalID: proposalID,
 		},
 	}
-	if value == VoteValueYes {
-		req.Vote.Value = types.Vote_VALUE_YES
-	}
 	resp, err := r.tradingClient.PrepareVote(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	gqlValue, err := convertVoteValueFromProto(resp.Vote.Value)
 	if err != nil {
 		return nil, err
 	}
@@ -1571,7 +1530,7 @@ func (r *myMutationResolver) PrepareVote(ctx context.Context, value VoteValue, p
 		Vote: &ProposalVote{
 			Vote: &Vote{
 				Party: party,
-				Value: VoteValueFromProto(resp.Vote.Value),
+				Value: gqlValue,
 			},
 			ProposalID: resp.Vote.ProposalID,
 		},
@@ -1607,7 +1566,7 @@ func (r *myMutationResolver) PrepareOrderAmend(ctx context.Context, id string, p
 		return nil, errors.New("invalid size, could not convert to unsigned int")
 	}
 
-	order.TimeInForce, err = parseOrderTimeInForce(tif)
+	order.TimeInForce, err = convertOrderTimeInForceToProto(tif)
 	if err != nil {
 		r.log.Error("unable to parse time in force in order amend",
 			logging.Error(err))
@@ -1919,7 +1878,7 @@ func (r *mySubscriptionResolver) MarketDepth(ctx context.Context, market string)
 
 func (r *mySubscriptionResolver) Candles(ctx context.Context, market string, interval Interval) (<-chan *types.Candle, error) {
 
-	pinterval, err := convertInterval(interval)
+	pinterval, err := convertIntervalToProto(interval)
 	if err != nil {
 		r.log.Debug("invalid interval for candles subscriptions", logging.Error(err))
 	}
@@ -2109,18 +2068,7 @@ func (r *myAccountResolver) Market(ctx context.Context, acc *types.Account) (*Ma
 }
 
 func (r *myAccountResolver) Type(ctx context.Context, obj *types.Account) (AccountType, error) {
-	var t AccountType
-	switch obj.Type {
-	case types.AccountType_ACCOUNT_TYPE_MARGIN:
-		return AccountTypeMargin, nil
-	case types.AccountType_ACCOUNT_TYPE_GENERAL:
-		return AccountTypeGeneral, nil
-	case types.AccountType_ACCOUNT_TYPE_INSURANCE:
-		return AccountTypeInsurance, nil
-	case types.AccountType_ACCOUNT_TYPE_SETTLEMENT:
-		return AccountTypeSettlement, nil
-	}
-	return t, ErrUnknownAccountType
+	return convertAccountTypeFromProto(obj.Type)
 }
 
 // END: Account Resolver