+++ governance/networkparams.go
@@ -1,9 +1,10 @@
 package governance
 
 import (
-	"fmt"
 	"strconv"
 	"time"
+
+	"code.vegaprotocol.io/vega/logging"
 )
 
 const (
@@ -17,7 +18,10 @@ var (
 	MaxClose              = ""
 	MinEnact              = ""
 	MaxEnact              = ""
-	MinParticipationStake = ""
+	RequiredParticipation = ""
+	RequiredMajority      = ""
+	MinProposerBalance    = ""
+	MinVoterBalance       = ""
 )
 
 const (
@@ -30,72 +34,208 @@ const (
 	defaultMinEnact = 2 * day // must be >= minClose
 	// defaultMaxEnact is the hardcoded maximum enactment offset duration
 	defaultMaxEnact = 1 * year
-	// defaultMinParticipationStake is hardcoded minimum participation stake percent
-	defaultMinParticipationStake = 1
+	// defaultRequiredParticipation is hardcoded participation level required for any proposal to pass (from `0` to `1`)
+	defaultRequiredParticipation = 0.00001
+	// defaultRequiredMajority is hardcoded majority level required for any proposal to pass (from `0.5` to `1`)
+	defaultRequiredMajority = 0.66
+	// defaultProposerBalance is hardcoded minimum balance required for a party
+	// to be able to submit a new proposal (greater than `0` to `1`)
+	defaultMinProposerBalance = 0.00001
+	// defaultMinVoterBalance is hardcoded minimum balance required for a party
+	// to be able to cast a vote (greater than `0` to `1`).
+	defaultMinVoterBalance = 0.00001
 )
 
-// NetworkParameters stores governance network parameters
+// ProposalParameters stores proposal specific parameters
+type ProposalParameters struct {
+	MinClose              time.Duration
+	MaxClose              time.Duration
+	MinEnact              time.Duration
+	MaxEnact              time.Duration
+	RequiredParticipation float32
+	RequiredMajority      float32
+	MinProposerBalance    float32
+	MinVoterBalance       float32
+}
+
+// NetworkParameters stores network parameters per proposal type
 type NetworkParameters struct {
-	minClose              time.Duration
-	maxClose              time.Duration
-	minEnact              time.Duration
-	maxEnact              time.Duration
-	minParticipationStake uint64
+	newMarkets ProposalParameters
 }
 
 // DefaultNetworkParameters returns default, hardcoded, network parameters
-func DefaultNetworkParameters() *NetworkParameters {
-	var err error
-	result := &NetworkParameters{
-		minClose:              defaultMinClose,
-		maxClose:              defaultMaxClose,
-		minEnact:              defaultMinEnact,
-		maxEnact:              defaultMaxEnact,
-		minParticipationStake: defaultMinParticipationStake,
+func DefaultNetworkParameters(log *logging.Logger) *NetworkParameters {
+	return &NetworkParameters{
+		newMarkets: defaultNewMarketParameters(log),
 	}
+}
+
+func defaultNewMarketParameters(log *logging.Logger) ProposalParameters {
+	var err error
+	result := ProposalParameters{
+		MinClose:              defaultMinClose,
+		MaxClose:              defaultMaxClose,
+		MinEnact:              defaultMinEnact,
+		MaxEnact:              defaultMaxEnact,
+		RequiredParticipation: defaultRequiredParticipation,
+		RequiredMajority:      defaultRequiredMajority,
+		MinProposerBalance:    defaultMinProposerBalance,
+		MinVoterBalance:       defaultMinVoterBalance,
+	}
+
 	if len(MinClose) > 0 {
-		result.minClose, err = time.ParseDuration(MinClose)
+		result.MinClose, err = time.ParseDuration(MinClose)
 		if err != nil {
-			panic(fmt.Sprintf("Failed to parse time duration: %s", MinClose))
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("MinClose", MinClose),
+				logging.Error(err),
+			)
 		}
 	}
 	if len(MaxClose) > 0 {
-		result.maxClose, err = time.ParseDuration(MaxClose)
+		result.MaxClose, err = time.ParseDuration(MaxClose)
 		if err != nil {
-			panic(fmt.Sprintf("Failed to parse time duration: %s", MaxClose))
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("MaxClose", MaxClose),
+				logging.Error(err),
+			)
 		}
 	}
 	if len(MinEnact) > 0 {
-		result.minEnact, err = time.ParseDuration(MinEnact)
+		result.MinEnact, err = time.ParseDuration(MinEnact)
 		if err != nil {
-			panic(fmt.Sprintf("Failed to parse time duration: %s", MinEnact))
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("MinEnact", MinEnact),
+				logging.Error(err),
+			)
 		}
 	}
 	if len(MaxEnact) > 0 {
-		result.maxEnact, err = time.ParseDuration(MaxEnact)
+		result.MaxEnact, err = time.ParseDuration(MaxEnact)
 		if err != nil {
-			panic(fmt.Sprintf("Failed to parse time duration: %s", MaxEnact))
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("MaxEnact", MaxEnact),
+				logging.Error(err),
+			)
 		}
 	}
-	if len(MinParticipationStake) > 0 {
-		result.minParticipationStake, err = strconv.ParseUint(MinParticipationStake, 10, 64)
+	if len(RequiredParticipation) > 0 {
+		levelValue, err := strconv.ParseFloat(RequiredParticipation, 32)
 		if err != nil {
-			panic(fmt.Sprintf("Failed to parse time integer: %s", MinParticipationStake))
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("RequiredParticipation", RequiredParticipation),
+				logging.Error(err),
+			)
 		}
-		if result.minParticipationStake > 100 {
-			panic(fmt.Sprintf("Invalid MinParticipationStake (over 100): %d", result.minParticipationStake))
+		if levelValue < 0 {
+			log.Fatal(
+				"New market network parameter is invalid (negative)",
+				logging.String("RequiredParticipation", RequiredParticipation),
+			)
 		}
+		if levelValue > 1 {
+			log.Fatal(
+				"New market network parameter is invalid (over 1)",
+				logging.String("RequiredParticipation", RequiredParticipation),
+			)
+		}
+		result.RequiredParticipation = float32(levelValue)
+	}
+	if len(RequiredMajority) > 0 {
+		levelValue, err := strconv.ParseFloat(RequiredMajority, 32)
+		if err != nil {
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("RequiredMajority", RequiredMajority),
+				logging.Error(err),
+			)
+		}
+		if levelValue < 0.5 {
+			log.Fatal(
+				"New market network parameter is invalid (less than 0.5)",
+				logging.String("RequiredMajority", RequiredMajority),
+			)
+		}
+		if levelValue > 1 {
+			log.Fatal(
+				"New market network parameter is invalid (over 1)",
+				logging.String("RequiredMajority", RequiredMajority),
+			)
+		}
+		result.RequiredMajority = float32(levelValue)
+	}
+	if len(MinProposerBalance) > 0 {
+		levelValue, err := strconv.ParseFloat(MinProposerBalance, 32)
+		if err != nil {
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("MinProposerBalance", MinProposerBalance),
+				logging.Error(err),
+			)
+		}
+		if levelValue <= 0 {
+			log.Fatal(
+				"New market network parameter is invalid (less or equal than 0)",
+				logging.String("MinProposerBalance", MinProposerBalance),
+			)
+		}
+		if levelValue > 1 {
+			log.Fatal(
+				"New market network parameter is invalid (over 1)",
+				logging.String("MinProposerBalance", MinProposerBalance),
+			)
+		}
+		result.MinProposerBalance = float32(levelValue)
+	}
+	if len(MinVoterBalance) > 0 {
+		levelValue, err := strconv.ParseFloat(MinVoterBalance, 32)
+		if err != nil {
+			log.Fatal(
+				"Failed to parse new market network parameter",
+				logging.String("MinVoterBalance", MinVoterBalance),
+				logging.Error(err),
+			)
+		}
+		if levelValue <= 0 {
+			log.Fatal(
+				"New market network parameter is invalid (less or equal than 0)",
+				logging.String("MinVoterBalance", MinVoterBalance),
+			)
+		}
+		if levelValue > 1 {
+			log.Fatal(
+				"New market network parameter is invalid (over 1)",
+				logging.String("MinVoterBalance", MinVoterBalance),
+			)
+		}
+		result.MinVoterBalance = float32(levelValue)
 	}
 
-	result.maxClose = max(result.maxClose, result.minClose) // maxClose must be >= minClose
-	result.minEnact = max(result.minEnact, result.minClose) // minEnact must be >= minClose
-	result.maxEnact = max(result.maxEnact, result.minEnact) // maxEnact must be >= minEnact
+	if result.MaxClose < result.MinClose {
+		log.Fatal(
+			"New market MaxClose network parameter is less than MinClose",
+			logging.String("MaxClose", result.MaxClose.String()),
+			logging.String("MinClose", result.MinClose.String()),
+		)
+	}
+	if result.MaxEnact < result.MinEnact {
+		log.Fatal(
+			"New market MaxEnact network parameter is less than MinEnact",
+			logging.String("MaxEnact", result.MaxEnact.String()),
+			logging.String("MinEnact", result.MinEnact.String()),
+		)
+	}
+	if result.MinEnact < result.MinClose {
+		log.Fatal(
+			"New market MinEnact network parameter is less than MinClose",
+			logging.String("MinEnact", result.MinEnact.String()),
+			logging.String("MinClose", result.MinClose.String()),
+		)
+	}
 	return result
-}
-
-func max(left, right time.Duration) time.Duration {
-	if left > right {
-		return left
-	}
-	return right
 }