+++ integration/core_test.go
@@ -136,7 +136,6 @@ func theMarket(mSetup *gherkin.DataTable) error {
 		mktsetup.party, // party-engine here!
 		mkt,
 		mktsetup.candles,
-		mktsetup.trades,
 		mktsetup.marginLevelsBuf,
 		NewSettlementStub(),
 		time.Now(),