+++ integration/setup_test.go
@@ -52,7 +52,6 @@ type marketTestSetup struct {
 	core            *execution.Market
 	party           *execution.Party
 	candles         *mocks.MockCandleBuf
-	trades          *tradeStub
 	accounts        *accStub
 	marginLevelsBuf *marginsStub
 	settle          *SettleStub
@@ -82,7 +81,6 @@ func getMarketTestSetup(market *proto.Market) *marketTestSetup {
 	lossBuf.EXPECT().Add(gomock.Any()).AnyTimes()
 	lossBuf.EXPECT().Flush().AnyTimes()
 	broker := NewBrokerStub()
-	trades := NewTradeStub()
 
 	// this can happen any number of times, just set the mock up to accept all of them
 	// Over time, these mocks will be replaced with stubs that store all elements to a map
@@ -101,7 +99,6 @@ func getMarketTestSetup(market *proto.Market) *marketTestSetup {
 		market:          market,
 		ctrl:            ctrl,
 		candles:         candles,
-		trades:          trades,
 		marginLevelsBuf: marginLevelsBuf,
 		settle:          NewSettlementStub(),
 		lossSoc:         lossBuf,
@@ -121,7 +118,6 @@ type executionTestSetup struct {
 	log             *logging.Logger
 	ctrl            *gomock.Controller
 	candles         *mocks.MockCandleBuf
-	trades          *tradeStub
 	markets         *mocks.MockMarketBuf
 	timesvc         *timeStub
 	marketdata      *mocks.MockMarketDataBuf
@@ -165,7 +161,6 @@ func getExecutionTestSetup(startTime time.Time, mkts []proto.Market) *executionT
 	execsetup.cfg.InsurancePoolInitialBalance = execsetup.InsurancePoolInitialBalance
 	execsetup.log = logging.NewTestLogger()
 	execsetup.candles = mocks.NewMockCandleBuf(ctrl)
-	execsetup.trades = NewTradeStub()
 	execsetup.settle = buffer.NewSettlement()
 	execsetup.markets = mocks.NewMockMarketBuf(ctrl)
 	execsetup.accs = map[string][]account{}
@@ -186,7 +181,7 @@ func getExecutionTestSetup(startTime time.Time, mkts []proto.Market) *executionT
 	execsetup.candles.EXPECT().AddTrade(gomock.Any()).AnyTimes()
 	execsetup.markets.EXPECT().Flush().AnyTimes().Return(nil)
 
-	execsetup.engine = execution.NewEngine(execsetup.log, execsetup.cfg, execsetup.timesvc, execsetup.trades, execsetup.candles, execsetup.markets, execsetup.marketdata, execsetup.marginLevelsBuf, execsetup.settle, execsetup.lossSoc, execsetup.proposal, execsetup.votes, mkts, execsetup.broker)
+	execsetup.engine = execution.NewEngine(execsetup.log, execsetup.cfg, execsetup.timesvc, execsetup.candles, execsetup.markets, execsetup.marketdata, execsetup.marginLevelsBuf, execsetup.settle, execsetup.lossSoc, execsetup.proposal, execsetup.votes, mkts, execsetup.broker)
 
 	// instanciate position plugin
 	execsetup.positionPlugin = plugins.NewPositions(execsetup.settle, execsetup.lossSoc)